+++ app/src/main/java/com/averycorp/prismtask/di/DatabaseModule.kt
@@ -3,6 +3,7 @@ package com.averycorp.prismtask.di
 import android.content.Context
 import androidx.room.Room
 import com.averycorp.prismtask.data.local.dao.AttachmentDao
+import com.averycorp.prismtask.data.local.dao.BatchUndoLogDao
 import com.averycorp.prismtask.data.local.dao.BoundaryRuleDao
 import com.averycorp.prismtask.data.local.dao.CalendarSyncDao
 import com.averycorp.prismtask.data.local.dao.CheckInLogDao
@@ -177,6 +178,9 @@ object DatabaseModule {
     @Provides
     fun provideMedicationTierStateDao(database: PrismTaskDatabase): MedicationTierStateDao =
         database.medicationTierStateDao()
+
+    @Provides
+    fun provideBatchUndoLogDao(database: PrismTaskDatabase): BatchUndoLogDao = database.batchUndoLogDao()
 }
 
 /**